Move polling function to services
