+++ bankid-frontend/src/views/AuthenticateView.vue
@@ -6,8 +6,8 @@
   import QrDisplay from '@/components/QrDisplay.vue';
   import QrInstructions from '@/components/QrInstructions.vue';
   import { PATHS } from '@/Redirects';
-  import { cancel, poll } from '@/Service';
-  import type { ApiResponse, ApiResponseStatus, RetryResponse, UiInformation } from '@/types';
+  import { cancel, polling } from '@/Service';
+  import type { ApiResponseStatus, UiInformation } from '@/types';
 
   const qrImage = ref('');
   const token = ref('');
@@ -22,47 +22,6 @@
   const showQrInstructions = computed(
     () => messageCode.value === 'bankid.msg.ext2' && props.uiInfo && props.uiInfo.displayQrHelp,
   );
-
-  function isApiResponse(obj: any): obj is ApiResponse {
-    return obj && 'status' in obj;
-  }
-
-  function isRetryResponse(obj: any): obj is RetryResponse {
-    return obj && 'retry' in obj;
-  }
-
-  const polling = () => {
-    poll(props.otherDevice).then((response) => {
-      if (isApiResponse(response)) {
-        responseStatus.value = response.status;
-        if (response.qrCode !== '') {
-          qrImage.value = response.qrCode;
-        }
-        if (response.status !== 'NOT_STARTED') {
-          qrImage.value = '';
-        }
-        token.value = response.autoStartToken;
-        messageCode.value = response.messageCode;
-
-        if (response.status === 'COMPLETE') {
-          window.location.href = PATHS.COMPLETE;
-        } else if (response.status === 'CANCEL') {
-          window.location.href = PATHS.CANCEL;
-        } else if (response.status === 'ERROR') {
-          qrImage.value = '';
-        }
-      }
-      if (isRetryResponse(response) && response.retry === true) {
-        /* Time is defined in seconds and setTimeout is in milliseconds */
-        window.setTimeout(() => polling(), parseInt(response.time) * 1000);
-      } else if (
-        isRetryResponse(response) ||
-        (isApiResponse(response) && (response.status === 'NOT_STARTED' || response.status === 'IN_PROGRESS'))
-      ) {
-        window.setTimeout(() => polling(), 500);
-      }
-    });
-  };
 
   const cancelRequest = async () => {
     await cancel();
@@ -85,7 +44,7 @@
   });
 
   onMounted(() => {
-    polling();
+    polling(props.otherDevice, qrImage, token, messageCode, responseStatus);
   });
 </script>
